# docs/swagger.yaml
basePath: /
definitions:
  model.Note:
    properties:
      color:
        type: string
      content:
        type: string
      created_at:
        type: string
      id:
        type: integer
      title:
        type: string
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
  model.User:
    properties:
      email:
        type: string
      id:
        type: integer
      nickname:
        type: string
      password:
        type: string
    type: object
  noteshandlers.DeleteResponse:
    properties:
      id:
        type: integer
    type: object
  userhandlers.AuthCheckResponse:
    properties:
      isAuthenticated:
        type: boolean
      user:
        additionalProperties: true
        type: object
    type: object
  userhandlers.DeleteResponse:
    properties:
      id:
        type: integer
    type: object
  userhandlers.LoginDTO:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  userhandlers.UserResponseDTO:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      nickname:
        type: string
      updated_at:
        type: string
    type: object
host: localhost:3001
info:
  contact: {}
  description: Simple notes service
  title: Notes API
  version: "1.0"
paths:
  /api/notes:
    post:
      consumes:
      - application/json
      description: создание Note
      parameters:
      - description: Создание Note
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/model.Note'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Note'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: NoteCreate
      tags:
      - Notes
  /api/notes/{id}:
    delete:
      consumes:
      - application/json
      description: удаление Note
      parameters:
      - description: ID заметки для удаления
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/noteshandlers.DeleteResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: NoteDelete
      tags:
      - Notes
    get:
      consumes:
      - application/json
      description: GetOneNote
      parameters:
      - description: ID пользователя для поиска
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Note'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: GetOneNote
      tags:
      - Notes
    patch:
      consumes:
      - application/json
      description: обновление Note
      parameters:
      - description: обновление User
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/model.Note'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: NoteUpdate
      tags:
      - Notes
  /api/users:
    post:
      consumes:
      - application/json
      description: создание User
      parameters:
      - description: Создание User
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/model.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/userhandlers.UserResponseDTO'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: UserCreate
      tags:
      - Users
  /api/users/{id}:
    delete:
      consumes:
      - application/json
      description: удаление User
      parameters:
      - description: ID пользователя для удаления
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/userhandlers.DeleteResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: UserDelete
      tags:
      - Users
    get:
      consumes:
      - application/json
      description: GetUserById
      parameters:
      - description: ID пользователя для поиска
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/userhandlers.UserResponseDTO'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: GetOneUser
      tags:
      - Users
    patch:
      consumes:
      - application/json
      description: обновление User
      parameters:
      - description: обновление User
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/model.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: UserUpdate
      tags:
      - Users
  /api/users/auth-check:
    get:
      consumes:
      - application/json
      description: Приходит кука с токеном, высылается 200, если токен ещё валиден
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/userhandlers.AuthCheckResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Проверка авторизации
      tags:
      - Auth
  /api/users/login:
    post:
      consumes:
      - application/json
      description: Вход стандарт email password, запись в httpOnly Cookies JWT, в
        body лежит user
      parameters:
      - description: Данные для логина
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/userhandlers.LoginDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          headers:
            Set-Cookie:
              description: JWT-токен
              type: string
          schema:
            $ref: '#/definitions/userhandlers.UserResponseDTO'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Логин
      tags:
      - Auth
  /api/users/logout:
    get:
      description: Простой get для logout. Пока только удаляет JWT из браузера, не
        ведёт blacklist
      responses:
        "204":
          description: No Content
      summary: Логаут
      tags:
      - Auth
swagger: "2.0"